+++ framework/devicedeps/darwin/cpu/meson.build
@@ -1,0 +1,9 @@
+# Copyright 2022 Intel Corporation.
+# SPDX-License-Identifier: Apache-2.0
+
+framework_files += \
+    files(
+        'cpu_affinity.cpp',
+        '../../generic' / device_type / 'msr.c',
+        '../../generic' / device_type / 'physicaladdress.c',
+    )
