+++ framework/devicedeps/freebsd/cpu/effective_cpu_freq.hpp
@@ -1,0 +1,1 @@
+#include "../../generic/cpu/effective_cpu_freq.hpp"
