+++ framework/devicedeps/freebsd/cpu/thermal_monitor.hpp
@@ -1,0 +1,1 @@
+#include "../../generic/cpu/thermal_monitor.hpp"
