+++ framework/devicedeps/generic/cpu/msr.c
