+++ framework/devicedeps/generic/cpu/physicaladdress.c
