+++ framework/devicedeps/generic/cpu/thermal_monitor.hpp
