+++ framework/devicedeps/linux/cpu/msr.c
