+++ framework/devicedeps/linux/cpu/thermal_monitor.hpp
