+++ framework/sysdeps/darwin/meson.build
@@ -3,11 +3,8 @@
 
 framework_files += \
     files(
-        'cpu_affinity.cpp',
         'malloc.cpp',
         'pthread_barrier.cpp',
         '../generic/kvm.c',
         '../generic/memfpt.c',
-        '../generic/msr.c',
-        '../generic/physicaladdress.c',
     )
