+++ framework/sysdeps/linux/meson.build
@@ -3,23 +3,18 @@
 
 if (host_machine.cpu_family() == 'x86_64')
     framework_files += files(
-        'interrupt_monitor.cpp',
         'kvm.c',
-        'msr.c',
     )
 else
     framework_files += files(
         '../generic/kvm.c',
-        '../generic/msr.c',
     )
 endif
 
 framework_files += \
     files(
-        'cpu_affinity.cpp',
         'malloc.cpp',
         'memfpt.cpp',
-        'physicaladdress.cpp',
     )
 
 if get_option('cpp_link_args').contains('-static')
