--- framework/sysdeps/unix/effective_cpu_freq.hpp
@@ -1,1 +1,0 @@
-#include "../generic/effective_cpu_freq.hpp"
