--- framework/sysdeps/windows/thermal_monitor.hpp
@@ -1,1 +1,0 @@
-#include "../generic/thermal_monitor.hpp"
